Improve Boursobank Gocardless transaction parsing (#5202)
* Fix Boursobank transfer parsing

As per the comments on https://github.com/actualbudget/actual/pull/4958#issuecomment-2988814739,
the Boursobank transfer parsing was not working correctly.

Indeed, the array returned by GoCardless for transfers is randomly
ordered. So we cannot rely on the first line to know its type.

To work around this, the code gets a bit more complex by:
- Checking the first line for known types (card, loan, atm withdrawal)
  and handling them accordingly.
- If it's not one of these, we iterate through all the lines by checking
  if the array contains a line with the transfer type.

* Add credit note type for BoursoBank

An `Avoir` is a refund made to the credit card.
Adds the proper payee / notes parsing for it.

* Improve Boursobank card transaction parsing

Some Boursobank transactions have an unknown number attached to the
payee name. Remove it from the payee name to ensure consistency across
transactions.

For instance, `CARTE 19/03/25 Github 4 CB*0494` (notice the `4`).

* Improve Boursobank payee name backslashes handling

After more testing, I found that the backslashes present are for the
pending transactions, and indicating the localization of the payment.

To keep the payee name consistent, remove what follows the backslashes.

For instance `PICARD SA 1234\\PARIS\\ FR` is for a pending transaction
whereas `PICARD SA 1234` is what we receive for a completed one.

* Set notes with date for Boursobank card transactions

@mistyque requested to see the data of the card transaction in this
comment: https://github.com/actualbudget/actual/pull/4958#issuecomment-2981459622

* Add 5202 release note

* [autofix.ci] apply automated fixes

* refactor loops and match

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>
Co-authored-by: Matt Fiddaman <github@m.fiddaman.uk>